I have replaced tangent generation code from the previous rev: Tangents are generated on the CPU using an algorithm that just takes the cross between the normal and an arbitrary vector (more on this below). This algorithm can be implemented entirely in the vertex shader, as can be seen from the commented lines of code in bump_vert.glsl.
Unlike the previous vert shader that generated tangents, these tangentslook OK on the sphere as well as on the ground. Note, though, that this only happens because the chosen cross vector works out alright on the particular topologies of each mesh. I cannot garauntee this vector will work on a wider range of meshes.