Builds and runs in Ubuntu Linux.
Windows stuff has been moved aside so as to not clutter the Linux build, but to still be there when needed on that platform. As a consequence, output is built in the bin/ directory to keep DLLs in one place and Linux users must install dependencies (GLEW, IL, SDL) themselves.